+(acct-{group, user}/greetd-0, eclass/{mozcoreconf-v{4, 5}}.eclass, net-im/zoom-5.1.418436.0628), ^(app-emulation/wine-vanilla-5.11 -> 5.12, app-text/calibre-4.17.0-r1 -> 4.20, app-text/poppler-0.89.0 -> 0.90.0, dev-db/sqlite-3.32.0 -> 3.32.3, dev-lang/python-{3.7.8, 3.8.3} -> {3.7.8-r1, 3.8.3-r1}, dev-libs/libuv-1.38.0-r1 -> 1.38.1, dev-util/{spirv-tools-2020.4_pre20200429 -> 2020.3, vulkan-headers-1.2.145 -> 1.2.146, vulkan-tools-1.2.141 -> 1.2.145}, media-libs/{vulkan-{layers, loader}-1.2.141 -> 1.2.145, net-fs/nfs-utils-2.4.3 -> 2.5.1, net-misc/{curl-7.71.0 -> 7.71.1, dhcpcd-9.1.3 -> 9.1.4}, sys-fs/btrfs-progs-5.6.1 -> 5.7, sys-kernel/dracut-049-r3 -> 050-r1, www-client/opera-69.0.3686.36 -> 69.0.3686.49), *(app-shells/gentoo-bashcomp-20190211, dev-db/postgresql-12.3, dev-lang/mono-6.6.0.161, dev-util/spirv-headers-1.5.3, eclass/{acct-user, ada, git-r3, kernel-2, mozcoreconf-v6, php-ext-source-r3, python-{any{-r1}, single-r1, utils-r1}, scons-utils}.eclass, net-dns/bind-tools-9.16.4, net-misc/{freerdp-2.1.2-r1, ntp-4.2.8_p15}, sys-fs/e2fsprogs-1.45.6, sys-libs/e2fsprogs-libs-1.45.6, www-client/firefox-78.0.1)

+++ sys-kernel/dracut/dracut-050-r1.ebuild
@@ -10,8 +10,8 @@ if [[ ${PV} == 9999 ]] ; then
 	EGIT_REPO_URI="https://github.com/dracutdevs/dracut"
 else
 	[[ "${PV}" = *_rc* ]] || \
-	KEYWORDS="~alpha amd64 arm ~arm64 ia64 ~mips ppc ppc64 ~sparc x86"
-	SRC_URI="https://github.com/dracutdevs/dracut/archive/${PV}.tar.gz -> ${P}.tar.gz"
+	KEYWORDS="~alpha ~amd64 ~arm ~arm64 ~ia64 ~mips ~ppc ~ppc64 ~sparc ~x86"
+	SRC_URI="https://www.kernel.org/pub/linux/utils/boot/${PN}/${P}.tar.xz"
 fi
 
 DESCRIPTION="Generic initramfs generation tool"
@@ -31,7 +31,7 @@ RDEPEND="
 	>=sys-apps/kmod-23[tools]
 	|| (
 		>=sys-apps/sysvinit-2.87-r3
-		sys-apps/openrc[sysv-utils,selinux?]
+		sys-apps/openrc[sysv-utils(-),selinux?]
 		sys-apps/systemd[sysv-utils]
 	)
 	>=sys-apps/util-linux-2.21
@@ -58,27 +58,17 @@ BDEPEND="
 	virtual/pkgconfig
 "
 
-DOCS=( AUTHORS HACKING NEWS README README.generic README.kernel README.modules
+DOCS=( AUTHORS HACKING NEWS README.md README.generic README.kernel README.modules
 	README.testsuite TODO )
 
 QA_MULTILIB_PATHS="usr/lib/dracut/.*"
 
 PATCHES=(
-	"${FILESDIR}"/048-dracut-install-simplify-ldd-parsing-logic.patch
-	"${FILESDIR}"/049-40network-Don-t-include-40network-by-default.patch
-	"${FILESDIR}"/049-remove-bashism-in-various-boot-scripts.patch
-	"${FILESDIR}"/049-network-manager-call-the-online-hook-for-connected-d.patch
-	"${FILESDIR}"/049-install-dracut-install.c-install-module-dependencies.patch
-	"${FILESDIR}"/049-install-string_hash_func-should-not-be-fed-with-NULL.patch
-	"${FILESDIR}"/049-dracut.sh-Fix-udevdir-detection.patch
-	"${FILESDIR}"/049-rngd-new-module-running-early-during-boot-to-help-ge.patch
-	"${FILESDIR}"/049-fs-lib-drop-a-bashism.patch
-	"${FILESDIR}"/049-network-manager-remove-useless-use-of-basename.patch
-	"${FILESDIR}"/049-move-setting-the-systemdutildir-variable-before-it-s.patch
-	"${FILESDIR}"/049-dracut-install-Support-the-compressed-firmware-files.patch
-	"${FILESDIR}"/049-crypt-create-locking-directory-run-cryptsetup.patch
-	"${FILESDIR}"/049-network-manager-fix-getting-of-ifname-from-the-sysfs.patch
-	"${FILESDIR}"/049-configure-find-cflags-and-libs-for-fts-on-musl.patch
+	"${FILESDIR}"/050-Makefile-merge-main-version-and-git-version-earlier.patch
+	"${FILESDIR}"/050-dracut.sh-don-t-call-fsfreeze-on-subvol-of-root-file.patch
+	"${FILESDIR}"/050-Makefile-fix-VERSION-again.patch
+	"${FILESDIR}"/050-busybox-module-fix.patch
+	"${FILESDIR}"/050-gentoo-ldconfig-paths.patch
 )
 
 src_configure() {
@@ -94,7 +84,7 @@ src_configure() {
 	echo ./configure "${myconf[@]}"
 	./configure "${myconf[@]}" || die
 
-	if [[ ${PV} != 9999 ]] ; then
+	if [[ ${PV} != 9999 && ! -f dracut-version.sh ]] ; then
 		# Source tarball from github doesn't include this file
 		echo "DRACUT_VERSION=${PV}" > dracut-version.sh || die
 	fi
